+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project partially adheres to [Semantic Versioning](https://semver.org/s
 
 ## 0.3.0 - ____
 
+- `[+]` Added support for including dates in generated agenda
 - `[*]` Dropped Apache License -- AgendaCreator is now protected by the MPL 2.0 copyleft
 
 ## 0.2.6 - 2019-04-18